+++ src/plugin/config.ts
@@ -1,20 +1,12 @@
 import type { Config } from "tailwindcss";
-import { Logs } from "./logs";
 
 export class FactoryConfig {
-  static async getTailwindConfig(configPath?: string): Promise<Config | undefined> {
-    if (configPath) {
-      return import(configPath)
-        .then((res) => {
-          return res.default;
-        })
-        .catch((err) => {
-          Logs.info(configPath);
-          Logs.error("Tailwind config not found");
-          Logs.error(err);
-          return {};
-        });
+  static async getTailwindConfig(config: Promise<Config | undefined>): Promise<Config | undefined> {
+    if (config) {
+      const loadedConfig = await config;
+      return loadedConfig;
     }
+
     return;
   }
 }